fix a remaining thread safety issue in the random deprecation warning
