+++ hypothesis-python/docs/changelog.rst
@@ -24,7 +24,7 @@ Hypothesis 6.x
 6.135.30 - 2025-07-14
 ---------------------
 
-Fix a remaining thread-safety issue with the recursion limit warning Hypothesis issues when an outside caller sets ``sys.setrecursionlimit`` (see :ref:`v6.135.23`).
+Fix a remaining thread-safety issue with the recursion limit warning Hypothesis issues when an outside caller sets ``sys.setrecursionlimit`` (see :ref:`v6.135.23 <v6.135.23>`).
 
 .. _v6.135.29:
